+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/presencetracing/organizer/list/TraceLocationsFragment.kt
@@ -14,10 +14,10 @@ import com.google.android.material.transition.MaterialSharedAxis
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.TraceLocationOrganizerTraceLocationsListFragmentBinding
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.TraceLocation
+import de.rki.coronawarnapp.ui.dialog.displayDialog
 import de.rki.coronawarnapp.ui.presencetracing.attendee.checkins.CheckInsFragment
 import de.rki.coronawarnapp.ui.presencetracing.organizer.category.adapter.category.traceLocationCategories
 import de.rki.coronawarnapp.ui.presencetracing.organizer.details.QrCodeDetailFragmentArgs
-import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.di.AutoInject
 import de.rki.coronawarnapp.util.list.setupSwipe
 import de.rki.coronawarnapp.util.lists.decorations.TopBottomPaddingDecorator
@@ -175,19 +175,13 @@ class TraceLocationsFragment : Fragment(R.layout.trace_location_organizer_trace_
         }
     }
 
-    private fun showDeleteAllDialog() {
-        val deleteAllDialog = DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.trace_location_organiser_list_delete_all_popup_title,
-            R.string.trace_location_organiser_list_delete_all_popup_message,
-            R.string.trace_location_organiser_list_delete_all_popup_positive_button,
-            R.string.trace_location_organiser_list_delete_all_popup_negative_button,
-            positiveButtonFunction = {
-                viewModel.deleteAllTraceLocations()
-            },
-            isDeleteDialog = true
-        )
-        DialogHelper.showDialog(deleteAllDialog)
+    private fun showDeleteAllDialog() = displayDialog(isDeleteDialog = true) {
+        setTitle(R.string.trace_location_organiser_list_delete_all_popup_title)
+        setMessage(R.string.trace_location_organiser_list_delete_all_popup_message)
+        setPositiveButton(R.string.trace_location_organiser_list_delete_all_popup_positive_button) { _, _ ->
+            viewModel.deleteAllTraceLocations()
+        }
+        setNegativeButton(R.string.trace_location_organiser_list_delete_all_popup_negative_button) { _, _ -> }
     }
 
     private fun openCreateEventFragment(traceLocation: TraceLocation) {
@@ -205,19 +199,16 @@ class TraceLocationsFragment : Fragment(R.layout.trace_location_organizer_trace_
     }
 
     private fun showDeleteSingleDialog(traceLocation: TraceLocation, position: Int?) =
-        DialogHelper.showDialog(
-            DialogHelper.DialogInstance(
-                requireActivity(),
-                R.string.trace_location_organiser_list_delete_single_popup_title,
-                R.string.trace_location_organiser_list_delete_single_popup_message,
-                R.string.trace_location_organiser_list_delete_all_popup_positive_button,
-                R.string.trace_location_organiser_list_delete_all_popup_negative_button,
-                positiveButtonFunction = { viewModel.deleteSingleTraceLocation(traceLocation) },
-                negativeButtonFunction = { position?.let { traceLocationsAdapter.notifyItemChanged(position) } },
-                cancelFunction = { position?.let { traceLocationsAdapter.notifyItemChanged(position) } },
-                isDeleteDialog = true
-            )
-        )
+        displayDialog(isDeleteDialog = true) {
+            setTitle(R.string.trace_location_organiser_list_delete_single_popup_title)
+            setMessage(R.string.trace_location_organiser_list_delete_single_popup_message)
+            setPositiveButton(R.string.trace_location_organiser_list_delete_all_popup_positive_button) { _, _ ->
+                viewModel.deleteSingleTraceLocation(traceLocation)
+            }
+            setNegativeButton(R.string.trace_location_organiser_list_delete_all_popup_negative_button) { _, _ ->
+                position?.let { traceLocationsAdapter.notifyItemChanged(it) }
+            }
+        }
 
     private fun onScrollChange(extend: Boolean) =
         with(binding.qrCodeFab) {
